fixture tests

+++ src/main/java/org/pwr/lotnisko/repository/ReservationRepository.java
@@ -4,7 +4,7 @@ import org.pwr.lotnisko.model.Reservation;
 
 public interface ReservationRepository
 {
-    int addReservation(Reservation reservation);
+    long addReservation(Reservation reservation);
 
     boolean removeReservation(Reservation reservation);
